Implementing invaders shooting

+++ src/main/java/spacewars/model/game/elements/invaders/Invader.java
@@ -3,7 +3,7 @@ package spacewars.model.game.elements.invaders;
 import spacewars.model.game.elements.Element;
 
 public abstract class Invader extends Element {
-    public Invader(int x, int y) {
+    public Invader(double x, double y) {
         super(x, y);
     }
